+++ packages/thirdweb/src/exports/wallets/embedded.ts
@@ -11,4 +11,4 @@ export {
   authenticate,
   getUserEmail,
   type GetAuthenticatedUserParams,
-} from "../../wallets/in-app/core/authentication/index.js";
+} from "../../wallets/in-app/web/lib/auth/index.js";
