+++ packages/thirdweb/src/exports/wallets/in-app.ts
@@ -1,12 +1,15 @@
-export { inAppWallet } from "../../wallets/create-wallet.js";
+// --- KEEEP IN SYNC with exports/react-native/wallets/in-app ---
+
+export { inAppWallet } from "../../wallets/in-app/web/in-app.js";
 
 export {
   preAuthenticate,
   authenticate,
   getUserEmail,
   getUserPhoneNumber,
-  type GetAuthenticatedUserParams,
-} from "../../wallets/in-app/core/authentication/index.js";
+} from "../../wallets/in-app/web/lib/auth/index.js";
+
+export { type GetAuthenticatedUserParams } from "../../wallets/in-app/core/authentication/type.js";
 
 export { hasStoredPasskey } from "../../wallets/in-app/web/lib/auth/passkeys.js";
