+++ packages/thirdweb/src/wallets/create-wallet.ts
@@ -12,11 +12,10 @@ import type {
 } from "./wallet-types.js";
 
 import { trackConnect } from "../analytics/track.js";
-import type { ThirdwebClient } from "../client/client.js";
 import { coinbaseWalletSDK } from "./coinbase/coinbase-wallet.js";
 import { getCoinbaseWebProvider } from "./coinbase/coinbaseSDKWallet.js";
 import { COINBASE } from "./constants.js";
-import { logoutAuthenticatedUser } from "./in-app/core/authentication/index.js";
+import { inAppWallet } from "./in-app/web/in-app.js";
 import { smartWallet } from "./smart/smart-wallet.js";
 import type { WCConnectOptions } from "./wallet-connect/types.js";
 import { createWalletEmitter } from "./wallet-emitter.js";
@@ -289,148 +288,4 @@ export function createWallet<const ID extends WalletId>(
  */
 export function walletConnect() {
   return createWallet("walletConnect");
-}
-
-/**
- * Creates an in-app wallet.
- * @param createOptions - configuration options
- * @returns The created in-app wallet.
- * @example
- * ```ts
- * import { inAppWallet } from "thirdweb/wallets";
- *
- * const wallet = inAppWallet();
- *
- * const account = await wallet.connect({
- *   client,
- *   chain,
- *   strategy: "google",
- * });
- * ```
- *
- * Enable smart accounts and sponsor gas for your users:
- * ```ts
- * import { inAppWallet } from "thirdweb/wallets";
- * const wallet = inAppWallet({
- *  smartAccount: {
- *   chain: sepolia,
- *   sponsorGas: true,
- * },
- * });
- * ```
- *
- * Specify a logo for your login page
- * ```ts
- * import { inAppWallet } from "thirdweb/wallets";
- * const wallet = inAppWallet({
- *  metadata: {
- *   image: {
- *    src: "https://example.com/logo.png",
- *    alt: "My logo",
- *    width: 100,
- *    height: 100,
- *   },
- *  },
- * });
- * ```
- *
- * Hide the ability to export the private key within the Connect Modal
- * ```ts
- * import { inAppWallet } from "thirdweb/wallets";
- * const wallet = inAppWallet({
- *  hidePrivateKeyExport: true
- * });
- * ```
- * @wallet
- */
-export function inAppWallet(
-  createOptions?: CreateWalletArgs<"inApp">[1],
-): Wallet<"inApp"> {
-  const emitter = createWalletEmitter<"inApp">();
-  let account: Account | undefined = undefined;
-  let chain: Chain | undefined = undefined;
-  let client: ThirdwebClient | undefined;
-
-  return {
-    id: "inApp",
-    subscribe: emitter.subscribe,
-    getChain: () => chain,
-    getConfig: () => createOptions,
-    getAccount: () => account,
-    autoConnect: async (options) => {
-      const { autoConnectInAppWallet } = await import(
-        "./in-app/core/wallet/index.js"
-      );
-
-      const [connectedAccount, connectedChain] = await autoConnectInAppWallet(
-        options,
-        createOptions,
-      );
-      // set the states
-      client = options.client;
-      account = connectedAccount;
-      chain = connectedChain;
-      trackConnect({
-        client: options.client,
-        walletType: "inApp",
-        walletAddress: account.address,
-      });
-      // return only the account
-      return account;
-    },
-    connect: async (options) => {
-      const { connectInAppWallet } = await import(
-        "./in-app/core/wallet/index.js"
-      );
-
-      const [connectedAccount, connectedChain] = await connectInAppWallet(
-        options,
-        createOptions,
-      );
-      // set the states
-      client = options.client;
-      account = connectedAccount;
-      chain = connectedChain;
-      trackConnect({
-        client: options.client,
-        walletType: "inApp",
-        walletAddress: account.address,
-      });
-      // return only the account
-      return account;
-    },
-    disconnect: async () => {
-      // If no client is assigned, we should be fine just unsetting the states
-      if (client) {
-        const result = await logoutAuthenticatedUser({ client });
-        if (!result.success) {
-          throw new Error("Failed to logout");
-        }
-      }
-      account = undefined;
-      chain = undefined;
-      emitter.emit("disconnect", undefined);
-    },
-    switchChain: async (newChain) => {
-      if (createOptions?.smartAccount && client && account) {
-        // if account abstraction is enabled, reconnect to smart account on the new chain
-        const { autoConnectInAppWallet } = await import(
-          "./in-app/core/wallet/index.js"
-        );
-        const [connectedAccount, connectedChain] = await autoConnectInAppWallet(
-          {
-            chain: newChain,
-            client,
-          },
-          createOptions,
-        );
-        account = connectedAccount;
-        chain = connectedChain;
-      } else {
-        // if not, simply set the new chain
-        chain = newChain;
-      }
-      emitter.emit("chainChanged", newChain);
-    },
-  };
 }
